Set it so the __MACOSX folder is deleted on all platforms. This folder must exist from compressing the zip in OS X, not depending on the platform it is uncompressed on as originally thought.

+++ Source/MainComponent.cpp
@@ -446,10 +446,9 @@ void MainContentComponent::run()
         // ==== Delete downloaded files ====
         updateDirectory.deleteRecursively();
         
-        #if JUCE_MAC
         File deleteMe = alphaLiveDirectory.getChildFile("__MACOSX");
-        deleteMe.deleteRecursively();
-        #endif 
+		if (deleteMe.exists())
+			deleteMe.deleteRecursively();
         
         #if JUCE_MAC
         File appFile (alphaLiveDirectory.getFullPathName() + File::separatorString + "AlphaLive.app");
